authorization/requirements.txt
architect==0.6.0
flask==2.2.2
flask-marshmallow==0.14.0
Flask-Limiter==2.8.1
Flask-Migrate==4.0.0
marshmallow-sqlalchemy==0.28.1
pbkdf2-reboot==1.0
flask_sqlalchemy==3.0.2
opentelemetry-api==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-instrumentation-flask==0.36b0
opentelemetry-exporter-jaeger==1.15.0
psycopg2-binary==2.9.5
pydantic==1.10.2
PyJWT==2.6.0
redis==4.3.5
requests==2.28.1
gevent==22.10.2
flasgger==0.9.5
click==8.1.3